Isolate timeframe types
