+++ src/models/timeframe/creatable-timeframe.ts
@@ -1,0 +1,12 @@
+/*************************************************************************
+ * Copyright 2020 Gravwell, Inc. All rights reserved.
+ * Contact: <legal@gravwell.io>
+ *
+ * This software may be modified and distributed under the terms of the
+ * MIT license. See the LICENSE file for details.
+ **************************************************************************/
+
+import { PartialProps } from '../../functions/utils';
+import { Timeframe } from './timeframe';
+
+export type CreatableTimeframe = PartialProps<Timeframe, 'start' | 'end'>;
