+++ src/models/timeframe/raw-timeframe.ts
@@ -1,0 +1,14 @@
+/*************************************************************************
+ * Copyright 2020 Gravwell, Inc. All rights reserved.
+ * Contact: <legal@gravwell.io>
+ *
+ * This software may be modified and distributed under the terms of the
+ * MIT license. See the LICENSE file for details.
+ **************************************************************************/
+
+export interface RawTimeframe {
+	durationString: string | null;
+	timeframe: string;
+	start?: string | null; // Timestamp
+	end?: string | null; // Timestamp
+}
