+++ src/models/timeframe/timeframe.ts
@@ -1,0 +1,14 @@
+/*************************************************************************
+ * Copyright 2020 Gravwell, Inc. All rights reserved.
+ * Contact: <legal@gravwell.io>
+ *
+ * This software may be modified and distributed under the terms of the
+ * MIT license. See the LICENSE file for details.
+ **************************************************************************/
+
+export interface Timeframe {
+	durationString: string | null;
+	timeframe: string;
+	start: Date | null;
+	end: Date | null;
+}
